# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "This is Wasilny Api Documentation",
        "version": "0.1"
    },
    "paths": {
...
        "/api/user/profile/getMyProfile": {
            "get": {
                "tags": [
                    "UserProfile"
                ],
                "summary": "جلب ملفي الشخصي",
                "description": "تقوم بجلب الملف الشخصي",
                "operationId": "profile/getMyProfile",
                "responses": {
                    "200": {
                        "description": "Success credentials response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "id": 2,
                                    "name": "Ahmed Nagem",
                                    "phone": "966536301031",
                                    "created_at": "2022-01-17T23:03:14.000000Z",
                                    "updated_at": "2022-06-20T07:45:15.000000Z"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/user/profile/updateMyProfile": {
            "post": {
                "tags": [
                    "UserProfile"
                ],
                "summary": "تحديث ملفي الشخصي",
                "description": "تقوم بتحديث الملف الشخصي",
                "operationId": "profile/updateMyProfile",
                "requestBody": {
                    "description": "جميع الخيارات إلزامية",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Ahmed Nagem"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "966536301031"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "502": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "FALSE"
                                        },
                                        "status": {
                                            "type": "int",
                                            "example": 502
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": {
                                                "message": "error message"
                                            }
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/user/update_firebase_token": {
            "post": {
                "tags": [
                    "UserProfile"
                ],
                "summary": "تحديث ملفي الشخصي",
                "description": "تقوم بتحديث الملف الشخصي",
                "operationId": "user/update_firebase_token",
                "requestBody": {
                    "description": "جميع الخيارات إلزامية",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firebase_token"
                                ],
                                "properties": {
                                    "firebase_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "asdSjfHE54jSFXjhdfdfwerweSHHDkjk5443"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "TRUE"
                                        },
                                        "status": {
                                            "type": "int",
                                            "example": 200
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": {
                                                "message": "Updated Successfuly"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}